requirements.txt
Flask==3.0.0
pyodbc==5.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.4
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
waitress==2.1.2
wfastcgi==3.0.0